requirements.txt
django==5.2.5
djangorestframework
djangorestframework-simplejwt
channels
channels_redis
celery
redis
psycopg2-binary
social-auth-app-django
python-dotenv
openai>=1.40.0
Pillow==10.4.0
dj-database-url
